:recycle: ISSUE-331 : 코드리뷰 반영
